refactor(test): split bench tests.rs god-file into tests/ subdir
