+++ homeboy.json
@@ -698,7 +698,6 @@
         "parallel-implementation::src/core/agent_task_prompts.rs::ParallelImplementation",
         "parallel-implementation::src/core/extension/trace/attach.rs::ParallelImplementation",
         "structural::src/core/agent_task_controller_service/tests.rs::HighItemCount",
-        "structural::src/commands/bench/tests.rs::GodFile",
         "structural::src/commands::DirectorySprawl",
         "structural::src/core/agent_task_controller_service/artifacts.rs::HighItemCount",
         "structural::src/core/agent_task_controller_service/tests.rs::GodFile",
@@ -717,7 +716,7 @@
         "structural::src/core/triage/tests.rs::GodFile",
         "structural::src/core::DirectorySprawl",
         "structural::tests/core_agnostic_test.rs::HighItemCount",
-        "test_quality::src/commands/bench/tests.rs::VacuousTest",
+        "test_quality::src/commands/bench/tests/scenario_run_tests.rs::VacuousTest",
         "test_quality::src/core/agent_task_scheduler/tests/outcome_tests.rs::VacuousTest",
         "test_quality::src/core/code_audit/duplication/tests.rs::VacuousTest",
         "test_quality::src/core/runner/execution/tests/policy.rs::VacuousTest",
